refactor(core): share WebAuthn option generators
